max-texture-size: Change variable to uint64_t rather than casting.
Every single usage of this variable cast it to uint64_t.  Clearly the
function parameter should be uint64_t, and we should have the compiler
automatically convert values when passed in.

+++ tests/texturing/max-texture-size.c
@@ -137,20 +137,17 @@ isValidTexSize(GLenum target, GLenum internalFormat, int sideLength)
 }
 
 static GLfloat *
-initTexData (GLenum target, int sideLength)
+initTexData(GLenum target, uint64_t sideLength)
 {
 	uint64_t nPixels;
 	if (target == GL_TEXTURE_1D)
-		nPixels = (uint64_t)(sideLength);
+		nPixels = sideLength;
 	else if (target == GL_TEXTURE_2D ||
 		 target == GL_TEXTURE_RECTANGLE ||
 		 target == GL_TEXTURE_CUBE_MAP)
-		nPixels = (uint64_t)(sideLength) *
-			  (uint64_t)(sideLength);
+		nPixels = sideLength * sideLength;
 	else if (target == GL_TEXTURE_3D)
-		nPixels = (uint64_t)(sideLength) *
-			  (uint64_t)(sideLength) *
-			  (uint64_t)(sideLength);
+		nPixels = sideLength * sideLength * sideLength;
 	else {
 		assert(!"Unexpected target");
 		return NULL;
